+++ Project/src/main/java/com/example/RMMBR/DemoApplication.java
@@ -1,12 +1,7 @@
-package com.example.demo;
+package com.example.RMMBR;
 
-import com.example.demo.auth.controllers.CustomTask;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.util.Calendar;
-import java.util.Timer;
-import java.util.concurrent.TimeUnit;
 
 @SpringBootApplication
 public class DemoApplication {
