+++ Project/src/main/java/com/example/RMMBR/WebApplicationConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.RMMBR;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
