+++ Project/src/main/java/com/example/RMMBR/auth/controllers/AuthenticationController.java
@@ -1,14 +1,12 @@
-package com.example.demo.auth.controllers;
+package com.example.RMMBR.auth.controllers;
 
-import com.example.demo.auth.data.UserRepository;
-import com.example.demo.auth.models.Reminder;
-import com.example.demo.auth.models.User;
-import com.example.demo.auth.models.dto.LoginFormDTO;
-import com.example.demo.auth.models.dto.RegisterFormDTO;
+import com.example.RMMBR.auth.data.UserRepository;
+import com.example.RMMBR.auth.models.User;
+import com.example.RMMBR.auth.models.dto.LoginFormDTO;
+import com.example.RMMBR.auth.models.dto.RegisterFormDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
@@ -20,8 +18,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Arrays;
-import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
 
 @RestController
