+++ Project/src/main/java/com/example/RMMBR/auth/controllers/ReminderController.java
@@ -1,9 +1,9 @@
-package com.example.demo.auth.controllers;
+package com.example.RMMBR.auth.controllers;
 
-import com.example.demo.auth.data.ReminderRepository;
-import com.example.demo.auth.models.Reminder;
-import com.example.demo.auth.models.User;
-import com.example.demo.auth.models.dto.ReminderFormDTO;
+import com.example.RMMBR.auth.data.ReminderRepository;
+import com.example.RMMBR.auth.models.Reminder;
+import com.example.RMMBR.auth.models.User;
+import com.example.RMMBR.auth.models.dto.ReminderFormDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,11 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
