+++ Project/src/main/java/com/example/RMMBR/auth/data/ReminderRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.auth.data;
+package com.example.RMMBR.auth.data;
 
-import com.example.demo.auth.models.Reminder;
+import com.example.RMMBR.auth.models.Reminder;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
