+++ Project/src/main/java/com/example/RMMBR/auth/models/Reminder.java
@@ -1,8 +1,7 @@
-package com.example.demo.auth.models;
+package com.example.RMMBR.auth.models;
 
 
 import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
