+++ Project/src/test/java/com/example/RMMBR/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.RMMBR;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
